% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_assumption.R
\name{anderson_darling_test}
\alias{anderson_darling_test}
\title{Anderson-Darling-Test}
\usage{
anderson_darling_test(X)
}
\arguments{
\item{X}{a single numeric vector of data values.}
}
\value{
An htest object is returned. The critical value and the p-value
are contained in the object.
}
\description{
Performs the Anderson-Darling-Test test for normality.
}
\details{
The Anderson-Darling tests the null hypothesis that a sample is
drawn from a population that follows a particular distribution.
For the Anderson-Darling test, the critical values depend on which
distribution is being tested against. This function works for
normal distributions
}
\examples{
\dontrun{
z <- rnorm(100)
olsdiagnosticR:::anderson_darling_test(X = z)

y <- rexp(100)
olsdiagnosticR:::anderson_darling_test(X = y)
}
}
\references{
Stephens, M. A. (1974). EDF Statistics for Goodness of Fit and Some
Comparisons, Journal of the American Statistical Association,
Vol. 69, pp. 730-737.

Stephens, M. A. (1976). Asymptotic Results for Goodness-of-Fit
Statistics with Unknown Parameters, Annals of Statistics, Vol. 4,
pp. 357-369.

Stephens, M. A. (1977). Goodness of Fit for the Extreme Value
Distribution, Biometrika, Vol. 64, pp. 583-588.

Stephens, M. A. (1977). Goodness of Fit with Special Reference
to Tests for Exponentiality , Technical Report No. 262,
Department of Statistics, Stanford University, Stanford, CA.

Stephens, M. A. (1979). Tests of Fit for the Logistic Distribution
Based on the Empirical Distribution Function, Biometrika,
Vol. 66, pp. 591-595.
}
