% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier.R
\name{influence.obs}
\alias{influence.obs}
\title{Plots the influence of observation on the regression line}
\usage{
\method{influence}{obs}(lm)
}
\arguments{
\item{lm}{linear model}
}
\value{
A plot
}
\description{
Plots the influence of observation on the regression line
}
