% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier.R
\name{influence.plot}
\alias{influence.plot}
\title{Plots the influence of observation on the regression line}
\usage{
\method{influence}{plot}(r, h, d)
}
\arguments{
\item{lm}{linear model}
}
\value{
A plot
}
\description{
Plots the influence of observation on the regression line
}
