% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier.R
\name{influence_plot}
\alias{influence_plot}
\title{Leverage values vs. Studentized residuals}
\usage{
influence_plot(influence_obs)
}
\arguments{
\item{influence_obs}{a list with following values:
leverage.value, cooks.distance, studentized.residuals}
}
\value{
ggplot-object
}
\description{
Bubble-Plot of the Leverage values vs. Studentized residuals with the
scale of the bubble by the cook's distance
}
\details{
Further information about Studentized residuals, Leverage values or
cook's distances,  can be found in the documentation
\code{\link{influence_observation}}
}
\examples{
\dontrun{
set.seed(167)
X <- data.frame(matrix(rnorm(1000), nrow = 100))
resid <- rnorm(100)
inf_obs <- olsdiagnosticR:::influence_observation(X = X, e = error)
olsdiagnosticR:::influence_plot(influence_obs = inf_obs)
}

}
