% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_assumption.R
\name{jarque_bera}
\alias{jarque_bera}
\title{Jarque-Bera-Test}
\usage{
jarque_bera(X)
}
\arguments{
\item{X}{a single numeric vector of data values.}
}
\value{
An htest object is returned. The critical value and the p-value
are contained in the object.
}
\description{
Performs the Jarque-Bera test for normality.
}
\details{
The Jarque-Bera test statistic tests the null that the data is normally
distributed against an alternative that the data follow some other distribution.
The test statistic is based on two moments of the data, the skewness, and the kurtosis,
and has an asymptotic chi^2 distribution.
The test statistic is defined:

   JB = n(S^2/6+(K-3)^2/24)

where n is the number of data points, S is the sample skewness,
and K is the sample kurtosis of the data.
}
\examples{
\dontrun{
z <- rnorm(100)
olsdiagnosticR:::jarque_bera(X = z)

y <- rexp(100)
olsdiagnosticR:::jarque_bera(X = y)
}
}
\references{
Jarque, C. M. and Bera, A. K. (1980). Efficient test for normality,
homoscedasticity and serial independence of residuals.
Economic Letters, 6(3), pp. 255-259.
}
