% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olsdiagnostic.R
\name{olsdiagnostic}
\alias{olsdiagnostic}
\title{olsdiagnosticR}
\usage{
olsdiagnostic(object, normtest = "Shapiro-Wilk", pval = 0.05)
}
\arguments{
\item{object}{linear regression model from class 'lm'}

\item{normtest}{the type of normality test. Possible test are

'Shapiro-Wilk', 'Jarque-Bera' or 'Anderson-Darling'}

\item{pval}{p-value for the hypothesis test. Default is 0.05}
}
\value{
A olsdiagnosticR class with following entries:


 \strong{Normality_Tests:}

 This could be the Shapiro-Wilk test, Jarque-Bera test,
 Anderson-Darling test or Cramer-von-mises test.
 Further information can be found in the documentation:

 \code{\link{Shapiro_Wilk_test}}

 \code{\link{jarque_bera}}

 \code{\link{anderson_darling_test}}


\strong{Outlier:}

Possible.Outlier:
a data.frame with possible outlier

Cooks.Distance:
a vector with the cook's distance values

Hat.Values
a vector with the lavarage values

Studentized.Residuals
a vector with the studentized residuals

Standardized.Residuals
a vector with the standardized residuals

Further information can for these values be found in the
documentation:

\code{\link{influence_observation}}


\strong{Multicollinearity:}

VIF:
a vector of the variance inflation factors
Further information can be found in the
documentation:

\code{\link{VIF}}

Contain.Columns:
A string with the columns that would be avoid multicollinearity
in the data


\strong{Heteroskedasticity:}

A Breusch-Pagan 'htest' object.

Further information can be found in the
documentation:

\code{\link{bp_test}}


\strong{Plots:}

QQ_Plot: \code{\link{qq_plot}}

Residual_Hist: \code{\link{resid_hist}}

Box_Plot_X: \code{\link{box_plot_x}}

Regressor_hist: \code{\link{hist_x}}

Cooks_Distance_Plot: \code{\link{plot_cd}}

Bubble_Plot: \code{\link{influence_plot}}

Spread_Level_Plot: \code{\link{influence_plot}}
}
\description{
Regression diagnostics for a linar regression model.
}
\details{
olsdiagnostic was developed for the statistical consulting of the
University of Göttingen to assist students in checking common
assumptions of linear regression quickly and easily. It also
provides students with an automated recommendation for the further
procedure if an assumption is not fulfilled.
}
\examples{
\dontrun{
e_norm <- rnorm(n)
e_exp <- rexp(n)

x1 = rnorm(n = n, mean = 80, sd = 10)
x2 = rnorm(n = n, mean = 70, sd = 5)
x3 = 2 * x1 + 4 * x2 + rnorm(n, mean = 20, sd = 5)

y_n = 3 + x1 + x2 + e_norm
y_m_e = 4 + x1 + x2 + x3 + e_exp

mod_n <- lm(y_n ~ x1 + x2)
mod_m_e <- lm(y_m_e ~ x1 + x2 + x3)

olsdiagnosticR::olsdiagnostic(mod_n)
olsdiagnosticR::olsdiagnostic(mod_m_e, normtest = 'Jarque-Bera')
}

}
