% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier.R
\name{plot.cd}
\alias{plot.cd}
\title{Visualisation of the cook's distances}
\usage{
\method{plot}{cd}(cd, p)
}
\arguments{
\item{cd}{Cook's distance}
}
\value{
Barplot
}
\description{
Visualisation of the cook's distances
}
