% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier.R
\name{plot_cd}
\alias{plot_cd}
\title{Visualisation of the cook's distances}
\usage{
plot_cd(influence_obs)
}
\arguments{
\item{influence_obs}{a list with following values:
leverage.value, cooks.distance, studentized.residuals}
}
\value{
qqplot-object
}
\description{
A barplot for cook's distances values
}
\details{
Further information about cook's distances can be found in the
documentation \code{\link{influence_observation}}
}
\examples{
\dontrun{
X <- data.frame(matrix(rnorm(1000), nrow = 100))
resid <- rnorm(100)
inf_obs <- olsdiagnosticR:::influence_observation(X = X, e = error)
olsdiagnosticR:::plot_cd(influence_obs = inf_obs)
}

}
\references{
Cook, R. D. and Weisberg, S. (1984) Residuals and Influence in
Regression.

Wiley. Fox, J. (1997) Applied Regression, Linear Models, and
Related Methods.

Sage. Williams, D. A. (1987) Generalized linear model diagnostics
using the deviance and single case deletions. Applied Statistics
36, 181--191.
}
