% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_assumption.R
\name{qq_plot}
\alias{qq_plot}
\title{QQ-Plot of the residuals}
\usage{
qq_plot(influence_obs)
}
\arguments{
\item{influence_obs}{a list with following values:
leverage.value, cooks.distance, studentized.residuals}
}
\value{
qqplot-object
}
\description{
Plots  studentized residuals from a linear model, against
theoretical quantiles of a comparison distribution.
}
\details{
Draws theoretical quantile-comparison plots for variables and for
studentized residuals from a linear model. A comparison line is
drawn on the plot either through the quartiles of the two
distributions, or by robust regression.
}
\examples{
\dontrun{
X <- data.frame(matrix(rnorm(1000), nrow = 100))
resid <- rnorm(100)
inf_obs <- olsdiagnosticR:::influence_observation(X = X, e = error)
olsdiagnosticR:::qq_plot(influence_obs = inf_obs)
}

}
\references{
Fox, J. 2016 Applied Regression Analysis and Generalized Linear Models, Third Edition. Sage.

Fox, J. and Weisberg, S. 2019 An R Companion to Applied Regression, Third Edition, Sage.
Atkinson, A. C. 1985 Plots, Transformations, and Regression. Oxford.
}
