% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_assumption.R
\name{resid_hist}
\alias{resid_hist}
\title{Histogramm of the residuals}
\usage{
resid_hist(X)
}
\arguments{
\item{X}{a single numeric vector of data values.}
}
\value{
ggplot-object.
}
\description{
This function generates a histogramm of the residuals with
the theoretical normal density.
}
\examples{
set.seed(1234)
X <- rnorm(200)
olsdiagnosticR:::resid_hist(X = X)

}
