% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homoskedasticity.R
\name{Spread_level_plot}
\alias{Spread_level_plot}
\title{Spread-Level Plot}
\usage{
Spread_level_plot(fitted.value, influence_obs)
}
\arguments{
\item{fitted.value}{a single numeric vector of data values}

\item{influence_obs}{a list with following values:
leverage.value, cooks.distance, studentized.residuals}
}
\value{
qqplot-object
}
\description{
Creates plots for examining the possible dependence of spread on
level, or an extension of these plots to the studentized residuals
from linear models.
}
\details{
Plots abs(studentized residuals) vs. fitted values.
}
\examples{
\dontrun{
y_hat <- rnorm(100)
X <- (matrix(rnorm(1000), nrow = 100))
resid <- rnorm(100)
inf_obs <- olsdiagnosticR:::influence_observation(X = X, e = error)
olsdiagnosticR:::Spread_level_plot(fitted.value = y_hat,
                                   influence_obs = inf_obs)
}


}
\references{
Fox, J. (1997) Applied Regression, Linear Models, and Related Methods.
Sage. Hoaglin, D. C., Mosteller, F. and Tukey, J. W. (Eds.) (1983)
Understanding Robust and Exploratory Data Analysis. Wiley.
}
