% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homoskedasticity.R
\name{bp_test}
\alias{bp_test}
\title{Breusch-Pagan Test}
\usage{
bp_test(e, X)
}
\arguments{
\item{e}{a vector with error terms}

\item{X}{a dataframe with k parameter and n observations}
}
\value{
An htest object is returned. The critical value and the p-value are
contained in the object.
}
\description{
Calculate a Breusch-Pegan test for heteroscedasticity
}
\details{
The Breusch-Pagan test fits a linear regression model to the
residuals of a linear regression model and rejects if too
much of the variance is explained by the additional explanatory
variables. Under H0 the test statistic of the Breusch-Pagan test
follows a chi-squared distribution with k degrees of freedom.
}
\examples{
 \dontrun{
## generate a regressor
X <- rep(c(1, 2), 50)
## generate homoskedastic errors
err1 <- rnorm(100)
## generate heteroskedastic errors
err2 <- rnorm(100, sd = x)

# perform the Breusch-Pagan test
olsdiagnosticR:::bp_test(err1, X)
olsdiagnosticR:::bp_test(err2, X)
}
}
\references{
T.S. Breusch & A.R. Pagan (1979), A Simple Test for Heteroscedasticity and Random Coefficient Variation.
Econometrica 47, 1287--1294

R. Koenker (1981), A Note on Studentizing a Test for Heteroscedasticity. Journal of Econometrics 17, 107--112.
}
