% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier.R
\name{hist_x}
\alias{hist_x}
\title{Histogram for each variable in design matrix}
\usage{
hist_x(X)
}
\arguments{
\item{X}{design matrix}
}
\value{
Returns a histogramm for each variablen in the design matrix
}
\description{
Histogram for each variable in design matrix
}
