% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier.R
\name{influence.observation}
\alias{influence.observation}
\title{Plots the cooks distance}
\usage{
\method{influence}{observation}(mod)
}
\arguments{
\item{lm}{a lm model}
}
\value{
A plot of the cooks distance
}
\description{
Plots the cooks distance
}
