% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier.R
\name{influence.plot}
\alias{influence.plot}
\title{Plots the influence of observation on the regression line}
\usage{
\code{isnormalr:::influence.plot(t, h, d)}
}
\arguments{
\item{t}{numeric vector of studentized residuals}

\item{h}{numeric vector of hat values}

\item{cd}{numeric vector of cook distance values}
}
\value{
A plot
}
\description{



}
\references{

}
