% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_assumption.R
\name{resid_hist}
\alias{resid_hist}
\title{Histogramm of the residuals}
\usage{
resid_hist(error)
}
\arguments{
\item{df}{Dataframe}
}
\value{
Plot a histogram of the residuals with a normal density
}
\description{
This function genertates a Histogramm of the residuals with a normal Density
}
\examples{
set.seed(1234)
X <- matrix(c(rnorm(200), rnorm(200)), ncol=2)
e <- rnorm(200)
b <- matrix(x(2, 5), ncol=1)
y <- X \%*\% b + e
resid_hist(lm(y ~ X))
}
