% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multikolonarity.R
\name{VIF}
\alias{VIF}
\title{Variance Inflation Factor}
\usage{
VIF(X)
}
\arguments{
\item{X}{Data Frame}
}
\value{
Returns the observations that VIF is larger then 2
}
\description{
Variance Inflation Factor
}
